tests/wpt/web-platform-tests/css/cssom/insertRule-syntax-error-01.html

<!DOCTYPE html>
<link rel="help" href="https://drafts.csswg.org/cssom-1/#insert-a-css-rule">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
p { color: green; }
</style>
<p id="test">This text should be green.</p>
<script>
    test(function() {
        assert_throws("SyntaxError", () => document.styleSheets[0].insertRule("p { color: red; } garbage", 1));
        assert_equals(getComputedStyle(document.getElementById("test")).color, "rgb(0, 128, 0)");
    }, "A syntax error in insertRule should throw and not affect the style of the page");
</script>
